progress parsing of publications
The structure of the publications.json is chosen such that it can be easily parsed using the Scala upickle library. For instance, every publication JSON object has the same fields. For fields that are optional at the Scala level, the JSON field has a null value if there is no value and otherwise a singleton list (upickle then automatically parses such fields into a Scala option).